+++ mycodo/config.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '7.3.1'
-ALEMBIC_VERSION = '27e1eca963ab'
+ALEMBIC_VERSION = 'f9ddbe510462'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.